Rename JSON to Value
We're going to be shifting the marketing focus for Argo away from being a JSON
parser and towards being a general purpose parser for transforming loosely
typed data models into strongly typed ones. As a part of this change, we can
generalize our types a bit, and use the name `Value` for our input values,
instead of `JSON`. This change is purely cosmetic, but will result in a more
general API, letting us break a bit from our ties to JSON.